test(start-checking): Reinstate invalid state check
Change name of "with invalid options" to "with invalid state", so as to
imply start-checking.js is to ensure the minimum requirements on the state
object are met before passing through to check().
Check that the state object is passed through to the state mock as-is if
those requirements are met.